Bump RestSharp from 106.15.0 to 107.3.0

+++ AdobeSign.UserManagement.Core/Clients/GroupClient.cs
@@ -36,7 +36,7 @@ namespace AdobeSign.UserManagement.Core.Clients
 
             if (pageSize != null)
             {
-                request.AddParameter("pageSize", pageSize);
+                request.AddParameter("pageSize", (int)pageSize);
             }
 
             if (!string.IsNullOrEmpty(cursor))
